Tic Tac  Toe Game Simulator
Create an application that simulates a game of tic-tac-toe. You could
use a 3x3 grid of Label controls to
show the X’s and O’s.
The application should use a two-dimensional int array to simulate the
game board in memory. When
the user clicks the ‘New Game’ button, the application should step
through the array, storing a random
number in the range of 0 to 1 in each element. Let the number zero
represent the letter O, and the
number 1 represent the letter X. The form should then be updated to
display the game board. The
application should display a message indicating whether player X won,
player O won, or if the game was
a tie.